Divide the code into several much smaller files.
No user-facing changes
